+++ src/main/java/org/deepsymmetry/beatlink/data/TrackPositionUpdate.java
@@ -23,11 +23,22 @@ public class TrackPositionUpdate {
     public final int beatNumber;
 
     /**
-     * If {@code true}, this was created in response to a beat packet, so we know exactly where the player was at that
-     * point. Otherwise, we infer position based on how long has elapsed since the previous beat packet, and the
-     * intervening playback pitch and direction.
+     * If {@code true}, this was created in response to a beat packet or precise position packet, so we know exactly
+     * where the player was at that point. Otherwise, we infer position based on how long has elapsed since
+     * the previous beat packet, and the intervening playback pitch and direction.
      */
     public final boolean definitive;
+
+    /**
+     * If {@code true}, we are receiving precise position packets from this player, so we know exactly where the
+     * player is even when it is not playing, or jumping or looping within beats.
+     */
+    public final boolean precise;
+
+    /**
+     * If {@code true}, this position update resulted from the receipt of a beat packet.
+     */
+    public final boolean fromBeat;
 
     /**
      * If {@code true}, the player reported that it was playing when the update was received.
@@ -51,7 +62,8 @@ public class TrackPositionUpdate {
     public final BeatGrid beatGrid;
 
     /**
-     * Constructor simply sets the fields of this immutable value class.
+     * Constructor simply sets the fields of this immutable value class. Backwards-compatible version which
+     * infers a {@code false} value for {@code precise}.
      *
      * @param timestamp when this update was received
      * @param milliseconds how far into the track has the player reached
@@ -64,6 +76,27 @@ public class TrackPositionUpdate {
      */
     public TrackPositionUpdate(long timestamp, long milliseconds, int beatNumber, boolean definitive,
                                boolean playing, double pitch, boolean reverse, BeatGrid beatGrid) {
+        this(timestamp, milliseconds, beatNumber, definitive, playing, pitch, reverse, beatGrid,
+                false, false);
+    }
+
+    /**
+     * Constructor simply sets the fields of this immutable value class.
+     *
+     * @param timestamp when this update was received
+     * @param milliseconds how far into the track has the player reached
+     * @param beatNumber the beat number that was reported (or incremented) by this update
+     * @param definitive indicates if this was based on a direct report of track position from the player (i.e. a beat)
+     * @param playing indicates whether the player was actively playing a track when this update was received
+     * @param pitch the playback pitch (where 1.0 is normal speed) when this update was received
+     * @param reverse indicates if the player was playing backwards when this update was received
+     * @param beatGrid the track beat grid that was used to calculate the update
+     * @param precise this position was derived from a high-precision position packet, directly or through a beat
+     * @param fromBeat this position was created in response to a beat packet
+     */
+    public TrackPositionUpdate(long timestamp, long milliseconds, int beatNumber, boolean definitive,
+                               boolean playing, double pitch, boolean reverse, BeatGrid beatGrid,
+                               boolean precise, boolean fromBeat) {
         this.timestamp = timestamp;
         this.milliseconds = milliseconds;
         this.beatNumber = beatNumber;
@@ -72,6 +105,8 @@ public class TrackPositionUpdate {
         this.pitch = pitch;
         this.reverse = reverse;
         this.beatGrid = beatGrid;
+        this.precise = precise;
+        this.fromBeat = fromBeat;
     }
 
     @Override
@@ -79,6 +114,6 @@ public class TrackPositionUpdate {
         return "TrackPositionUpdate[timestamp:" + timestamp + ", milliseconds:" + milliseconds +
                 ", beatNumber:" + beatNumber + ", definitive:" + definitive + ", playing:" + playing +
                 ", pitch:" + String.format("%.2f", pitch) + ", reverse:" + reverse +
-                ", beatGrid:" + beatGrid + "]";
+                ", beatGrid:" + beatGrid + ", precise:" + precise + "]";
     }
 }